Extract rightbar component into two different component between profile and home
