+++ src/components/rightbar-profile/RightbarProfile.scss
@@ -1,0 +1,280 @@
+.rightbar-profile {
+  // flex: 3;
+  width: 100%;
+  // height: calc(100vh - 50px);
+  // overflow-y: scroll;
+  // position: sticky;
+  // top: 50px;
+  // direction: rtl;
+
+  // &::-webkit-scrollbar {
+  //   width: 3px;
+  //   display: none;
+  // }
+
+  // &::-webkit-scrollbar-track {
+  //   background-color: #f1f1f1;
+  // }
+
+  // &::-webkit-scrollbar-thumb {
+  //   background-color: rgb(209, 209, 209);
+  //   border-radius: 5px;
+  // }
+
+  @media only screen and (min-width: 1024px) {
+    flex: 3;
+    height: calc(100vh - 50px);
+    overflow-y: scroll;
+    position: sticky;
+    top: 50px;
+    direction: rtl;
+
+    &::-webkit-scrollbar {
+      width: 3px;
+      display: none;
+    }
+
+    &::-webkit-scrollbar-track {
+      background-color: #f1f1f1;
+    }
+
+    &::-webkit-scrollbar-thumb {
+      background-color: rgb(209, 209, 209);
+      border-radius: 5px;
+    }
+  }
+
+  .rightbar-profile-wrapper {
+    display: flex;
+    flex-direction: column;
+    direction: ltr;
+    padding: 10px 10px 0px 10px;
+
+    @media only screen and (min-width: 1024px) {
+      padding: 10px 10px 0px 10px;
+    }
+
+    .rightbar-ad {
+      width: 100%;
+      border-radius: 10px;
+      margin: 0px 0px 10px 0px;
+      cursor: pointer;
+      -webkit-box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+      box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+    }
+
+    .follow-button {
+      width: 100%;
+      text-align: center;
+      border-radius: 5px;
+      cursor: pointer;
+      padding: 6px 0;
+      margin-bottom: 10px;
+      font-size: 18px;
+      font-weight: 600;
+      border-color: blue;
+      border-width: 2px;
+      border-style: solid;
+
+      &.active {
+        background-color: blue;
+        color: white;
+      }
+
+      &.not-active {
+        color: blue;
+      }
+
+      &.loading {
+        color: #2c2891;
+        cursor: not-allowed;
+      }
+    }
+
+    .rightbar-title {
+      margin-bottom: 20px;
+    }
+
+    .empty-state-user-online {
+      width: 100%;
+      background-color: white;
+      color: rgb(61, 61, 173);
+      text-align: center;
+      font-size: 12px;
+      font-weight: 600;
+      line-height: 16px;
+      padding: 8px;
+      border-radius: 6px;
+      margin-bottom: 12px;
+      -webkit-box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+      box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+    }
+
+    .rightbar-friend-list {
+      padding: 10px;
+      margin: 0;
+      list-style: none;
+      background: white;
+      border-radius: 10px;
+      -webkit-box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+      box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+      display: flex;
+      flex-direction: column;
+      gap: 15px;
+    }
+
+    .rightbar-title {
+      font-size: 14px;
+      font-weight: 500;
+      margin-bottom: 10px;
+    }
+
+    .rightbar-info {
+      background-color: white;
+      width: 100%;
+      border-radius: 10px;
+      font-size: 12px;
+      padding: 10px;
+      box-sizing: border-box;
+      -webkit-box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+      box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+
+      @media only screen and (min-width: 1024px) {
+        margin-bottom: 10px;
+        border-radius: 5px;
+      }
+
+      .rightbar-info-item {
+        margin-bottom: 6px;
+        display: flex;
+        flex-direction: column;
+        align-items: baseline;
+
+        .rightbar-info-key {
+          font-weight: 500;
+          margin-right: 4px;
+          color: #555;
+
+          .user-following-info {
+            width: fit-content;
+            font-size: 9px;
+            padding: 4px;
+            border-radius: 6px;
+            margin-left: 6px;
+            background-color: rgb(109, 109, 240);
+            color: white;
+          }
+        }
+
+        .rightbar-info-value {
+          font-weight: 300;
+          font-size: 12px;
+        }
+      }
+
+      .edit-profile-button {
+        border: none;
+        border-radius: 8px;
+        width: 100%;
+        background-color: #605beb;
+        color: white;
+        cursor: pointer;
+        text-align: center;
+        padding: 12px 0px;
+        margin-top: 10px;
+        box-sizing: border-box;
+        font-size: 11px;
+
+        @media only screen and (min-width: 1024px) {
+          padding: 8px 0px;
+          font-size: 12px;
+        }
+      }
+
+      .edit-profile-button:last-child {
+        margin-top: 14px;
+
+        @media only screen and (min-width: 1024px) {
+          margin-top: 12px;
+        }
+      }
+    }
+
+    .rightbar-followings-wrapper {
+      width: 100%;
+      height: fit-content;
+      padding: 10px;
+      background-color: white;
+      border-radius: 10px;
+      -webkit-box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+      box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+      margin-top: 10px;
+
+      @media only screen and (min-width: 1024px) {
+        border-radius: 5px;
+        margin-top: 0;
+        margin-bottom: 10px;
+      }
+
+      .rightbar-title-user-follower {
+        font-size: 14px;
+        font-weight: 500;
+        margin-bottom: 10px;
+      }
+
+      .rightbar-followings {
+        width: 100%;
+        display: flex;
+        flex-direction: row;
+        flex-wrap: wrap;
+        justify-content: space-between;
+        box-sizing: border-box;
+
+        .rightbar-following {
+          flex-shrink: 2;
+          display: flex;
+          flex-direction: column;
+          cursor: pointer;
+          width: 100%;
+          height: 100%;
+
+          @media only screen and (min-width: 320px) {
+            width: 30%;
+            height: 30%;
+          }
+
+          @media only screen and (min-width: 480px) {
+            width: 48%;
+            height: 48%;
+          }
+
+          .rightbar-following-img-wrapper {
+            width: 100%;
+            height: 100px;
+
+            .rightbar-following-img {
+              width: 100%;
+              height: 100%;
+              border-radius: 5px;
+              object-fit: fill;
+
+              @media only screen and (min-width: 320px) {
+                object-fit: cover;
+              }
+            }
+          }
+
+          .rightbar-following-name {
+            text-align: center;
+            font-size: 12px;
+            font-weight: 500;
+          }
+        }
+
+        .rightbar-following:last-child {
+          margin-bottom: 0px;
+        }
+      }
+    }
+  }
+}
